refactor and consolidate fixtures
